moves interpolate, docs
