+++ README.md
@@ -15,6 +15,8 @@ This is a library for writing and rendering SwiftUI-like views in Rust,
 primarily intended for use on `no_std` memory-constrained embedded systems.
 Floating point math is aggressively avoided.
 
+Nightly is required.
+
 ## Available render targets
 
 - `DrawTarget`: `embedded-graphics` displays.
@@ -29,8 +31,6 @@ Here's what an Apple-like toggle button component would look like,
 implemented with Buoyant:
 
 ```rust
-// Omitting the large number of imports...
-
 fn toggle_button(is_on: bool) -> impl Renderable<Rgb565, Renderables: EmbeddedGraphicsRender<Rgb565>> {
     let alignment = if is_on {
         HorizontalAlignment::Trailing
@@ -45,7 +45,7 @@ fn toggle_button(is_on: bool) -> impl Renderable<Rgb565, Renderables: EmbeddedGr
         Circle.foreground_color(Rgb565::WHITE).padding(2),
     ))
     .with_horizontal_alignment(alignment)
-    .frame(Some(50), Some(25), None, None)
+    .frame().with_width(50).with_height(25)
     .animated(Animation::Linear(Duration::from_millis(200)), is_on)
 }
 ```
