+++ src/lib.rs
@@ -8,7 +8,6 @@ extern crate std;
 pub mod environment;
 pub mod font;
 pub mod layout;
-pub mod pixel;
 pub mod primitives;
 pub mod render;
 pub mod render_target;
