+++ src/render_target/fixed_text_buffer.rs
@@ -3,6 +3,7 @@ use core::fmt::{Display, Formatter, Result};
 use crate::{primitives::Size, render::CharacterRenderTarget};
 
 /// A fixed size text buffer
+#[derive(Debug, Clone, PartialEq)]
 pub struct FixedTextBuffer<const W: usize, const H: usize> {
     pub text: [[char; W]; H],
 }
@@ -42,12 +43,14 @@ impl<const W: usize, const H: usize> CharacterRenderTarget for FixedTextBuffer<W
         character: char,
         _color: &Self::Color,
     ) {
+        #[allow(clippy::cast_sign_loss)]
         if point.x < W as i16 && point.y < H as i16 && point.x >= 0 && point.y >= 0 {
             self.text[point.y as usize][point.x as usize] = character;
         }
     }
 
     fn draw_color(&mut self, point: crate::primitives::Point, color: &Self::Color) {
+        #[allow(clippy::cast_sign_loss)]
         if point.x < W as i16 && point.y < H as i16 && point.x >= 0 && point.y >= 0 {
             self.text[point.y as usize][point.x as usize] = *color;
         }
