+++ src/view/modifier/fixed_size.rs
@@ -7,6 +7,7 @@ use crate::{
 
 /// Proposes ``ProposedDimension::Compact``, resulting in the child view rendering at its ideal
 /// size along the specified axis.
+#[derive(Debug, Clone)]
 pub struct FixedSize<T> {
     horizontal: bool,
     vertical: bool,
