+++ src/view/shape/circle.rs
@@ -5,6 +5,7 @@ use crate::{
     render::Renderable,
 };
 
+/// A circle primitive
 #[derive(Debug, Copy, Clone, PartialEq, Default)]
 pub struct Circle;
