+++ src/view/shape/rectangle.rs
@@ -7,10 +7,13 @@ use crate::{
 
 use super::RoundedRectangle;
 
+/// A rectangle
 #[derive(Debug, Copy, Clone, PartialEq, Default)]
 pub struct Rectangle;
 
 impl Rectangle {
+    /// Adds a corner radius to all corners. The radius will be capped to half the shorter
+    /// dimension
     #[must_use]
     pub fn corner_radius(self, radius: u16) -> RoundedRectangle {
         RoundedRectangle::new(radius)
